Rough draft of all files so far
This commit should never be used outside of testing purposes. It is very incomplete. Just a rough draft of everything I have now. Nothing is final and it contains many bugs.
The filter in the fetchRandomVideo.js file will return nothing/not enough songs if the difficulty is changed due to lack of songs and logic at the moment. This will change in the future